+++ ageoste/settings.py
@@ -29,7 +29,8 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'corsheaders',
     'user',
-    'product'
+    'product',
+    'order'
 ]
 
 MIDDLEWARE = [
