+++ ageoste/urls.py
@@ -3,4 +3,5 @@ from django.urls import path, include
 urlpatterns = [
     path('user', include('user.urls')),
     path('product', include('product.urls')),
+    path('order', include('order.urls'))
 ]
