+++ order/models.py
@@ -1,0 +1,30 @@
+ from django.db              import models
+ from django.core.validators import MinValueValidator, MaxValueValidator
+
+
+ class Order(models.Model):
+     user       = models.ForeignKey('User', on_delete=models.CASCADE)
+     created_at = models.DateTimeField(auto_now_add=True)
+     order_status = models.ForeignKey('OrderStatus', on_delete=models.CASCADE)
+
+     class Meta:
+         db_table = 'orders'
+
+
+ class OrderStatus(models.Model):
+     status = models.CharField(max_length=800)
+
+     class Meta:
+         db_table = 'order_statuses'
+
+
+ class CartItem(models.Model):
+     user    = models.ForeignKey('User', on_delete=models.CASCADE)
+     product = models.ForeignKey('product.Product', on_delete=models.CASCADE)
+     size    = models.ForeignKey('product.Size', on_delete=models.CASCADE)
+     color   = models.ForeignKey('product.Color',on_delete=models.CASCADE)
+     order   = models.ForeignKey('Order', on_delete=models.CASCADE)
+     count   = models.IntegerField()
+
+     class Meta:
+         db_table = 'cart_items'
