+++ user/models.py
@@ -53,32 +53,3 @@ class Membership(models.Model):
         db_table = 'memberships'
 
 
-class Order(models.Model):
-    user       = models.ForeignKey('User', on_delete=models.CASCADE)
-    created_at = models.DateTimeField(auto_now_add=True)
-
-
-    class Meta:
-        db_table = 'orders'
-
-
-class OrderStatus(models.Model):
-    order  = models.ForeignKey('Order', on_delete=models.CASCADE)
-    status = models.CharField(max_length=800)
-
-    class Meta:
-        db_table = 'order_statuses'
-
-
-class CartItem(models.Model):
-    user    = models.ForeignKey('User', on_delete=models.CASCADE)
-    product = models.ForeignKey('product.Product', on_delete=models.CASCADE)
-    size    = models.ForeignKey('product.Size', on_delete=models.CASCADE)
-    color   = models.ForeignKey('product.Color',on_delete=models.CASCADE)
-    order   = models.ForeignKey('Order', on_delete=models.CASCADE)
-    count   = models.IntegerField()
-
-    class Meta:
-        db_table = 'cart_items'
-
-
